requirements.txt
matplotlib
numpy
TensorFlow
tensorflow_hub
pillow
